Separate implementations from header file
